chore: Move Inquiry routes to Sell app
