+++ src/Apps/Sell/sellRoutes.tsx
@@ -1,6 +1,7 @@
 import loadable from "@loadable/component"
 import { RouteProps } from "System/Router/Route"
 import { getENV } from "Utils/getENV"
+import { RouteRenderArgs } from "found"
 import { graphql } from "react-relay"
 
 const IntroRoute = loadable(
@@ -158,6 +159,36 @@ const FAQApp = loadable(
   () => import(/* webpackChunkName: "consignBundle" */ "./Routes/FAQ/FAQApp"),
   {
     resolveComponent: component => component.FAQApp,
+  }
+)
+
+const ConsignmentInquiryApp = loadable(
+  () =>
+    import(
+      /* webpackChunkName: "consignBundle" */ "./Routes/ConsignmentInquiry/ConsignmentInquiry"
+    ),
+  {
+    resolveComponent: component =>
+      component.ConsignmentInquiryFragmentContainer,
+  }
+)
+const ConsignmentInquiryConfirmationApp = loadable(
+  () =>
+    import(
+      /* webpackChunkName: "consignBundle" */ "./Routes/ConsignmentInquiry/ConsignmentInquiryConfirmation"
+    ),
+  {
+    resolveComponent: component => component.ConsignmentInquiryConfirmation,
+  }
+)
+
+const ConsignmentInquiryContainer = loadable(
+  () =>
+    import(
+      /* webpackChunkName: "consignBundle" */ "./Routes/ConsignmentInquiry/ConsignmentInquiryContainer"
+    ),
+  {
+    resolveComponent: component => component.ConsignmentInquiryContainer,
   }
 )
 
@@ -505,6 +536,85 @@ export const sellRoutes: RouteProps[] = [
             prepareVariables: ({ id }) => {
               return { id }
             },
+          },
+        ],
+      },
+
+      // Inquiry Routes
+
+      {
+        path: "inquiry",
+        getComponent: () => ConsignmentInquiryContainer,
+        children: [
+          {
+            path: "/",
+            getComponent: () => ConsignmentInquiryApp,
+            layout: "ContainerOnly",
+            onClientSideRender: () => {
+              ConsignmentInquiryApp.preload()
+            },
+            query: graphql`
+              query sellRoutes_ConsignmentInquiryAppQuery {
+                me {
+                  ...ConsignmentInquiry_me
+                }
+                viewer {
+                  ...ConsignmentInquiry_viewer
+                }
+              }
+            `,
+            render: ({ Component, props }: RouteRenderArgs) => {
+              if (!(Component && props)) {
+                return undefined
+              }
+              return <Component {...props} />
+            },
+          },
+          {
+            path: "sent",
+            layout: "ContainerOnly",
+            getComponent: () => ConsignmentInquiryConfirmationApp,
+            onClientSideRender: () => {
+              ConsignmentInquiryConfirmationApp.preload()
+            },
+          },
+          {
+            path: "inquiry/:recipientEmail?",
+            getComponent: () => ConsignmentInquiryContainer,
+            children: [
+              {
+                path: "/",
+                getComponent: () => ConsignmentInquiryApp,
+                layout: "ContainerOnly",
+                onClientSideRender: () => {
+                  ConsignmentInquiryApp.preload()
+                },
+                query: graphql`
+                  query sellRoutes_ConsignmentInquiryWithRecipientEmailAppQuery {
+                    me {
+                      ...ConsignmentInquiry_me
+                    }
+                    viewer {
+                      ...ConsignmentInquiry_viewer
+                    }
+                  }
+                `,
+                render: ({ Component, props }: RouteRenderArgs) => {
+                  if (!(Component && props)) {
+                    return undefined
+                  }
+                  return <Component {...props} />
+                },
+              },
+              {
+                path: "sent",
+                layout: "ContainerOnly",
+                getComponent: () => ConsignmentInquiryConfirmationApp,
+                onClientSideRender: () => {
+                  ConsignmentInquiryConfirmationApp.preload()
+                },
+              },
+            ],
           },
         ],
       },
